The RE is updated for submission parsing and unit test are added. I also added FREE deal detection, but there are a few limitations for good reasons that I should mention to you guys as well:
Sometimes people put FREE information together with the title instead to
be with the price information. Our RE only detects "FREE" only if its
with the price. I did it this way because some deals literally contain
the word "FREE" in them, and there is no way to know if it is a part of
the deal or game title.

Not all "FREE" are %100 off. Sometimes, it is a deal along the lines of
"Buy X number of games get 1 FREE" OR "FREE during the weekends" etc...
They don't have a standard format.

What I can cover with the new RE:

When deals are separated with random/incorrect delimiters, we can still
parse.

When price information are separated with random/incorrect delimiters,
we can still parse.

Price contain '.' or ',' to separate dollar and cent value, also for GBP
and EURO.

% Euro $ GBP can come before or after the price/discount value.